% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_c_id}
\alias{check_c_id}
\title{check c_id arg to omop_ancestors() and omop_descendants()
NOT EXPORTED}
\usage{
check_c_id(c_id, anc_or_des)
}
\arguments{
\item{c_id}{arg to check : single omop concept_id or exact concept_name to get ancestors of}

\item{anc_or_des}{"ancestors" or "descendants"}
}
\description{
check c_id arg to omop_ancestors() and omop_descendants()
NOT EXPORTED
}
