% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{concept}
\alias{concept}
\title{concept table for selected OMOP vocabularies}
\format{
\code{dataframe}

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1326054 rows and 7 columns.
}
\source{
\url{https://athena.ohdsi.org/vocabulary/list}
}
\usage{
concept
}
\description{
concept table for selected OMOP vocabularies
}
\section{Slots}{

\describe{
\item{\code{data}}{OMOP concepts}
}}

\keyword{datasets}
