% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept_id.R
\name{concept_id}
\alias{concept_id}
\title{query a concept record by concept_id from omop concepts table}
\usage{
concept_id(c_id)
}
\arguments{
\item{c_id}{one concept_id to filter by}
}
\description{
query a concept record by concept_id from omop concepts table
}
\examples{
concept_id(3807321L)

}
