% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept_names.R
\name{concept_names}
\alias{concept_names}
\title{string search of concept_name in omop concepts table}
\usage{
concept_names(
  pattern,
  ignore_case = TRUE,
  negate = FALSE,
  c_ids = NULL,
  d_ids = NULL,
  v_ids = NULL,
  cc_ids = NULL,
  standard = NULL
)
}
\arguments{
\item{pattern}{string to search for or regex}

\item{ignore_case}{ignore case in string comparison, default TRUE}

\item{negate}{If TRUE, return non-matching elements, default FALSE}

\item{c_ids}{one or more concept_id to filter by, default NULL for all}

\item{d_ids}{one or more domain_id to filter by, default NULL for all}

\item{v_ids}{one or more vocabulary_id to filter by, default NULL for all}

\item{cc_ids}{one or more concept_class_id to filter by, default NULL for all}

\item{standard}{one or more standard_concept to filter by, default NULL for all, S,C}
}
\description{
string search of concept_name in omop concepts table
}
\examples{
concept_names("AJCC/UICC Stage")
chemo <- concept_names("chemotherapy", v_ids="LOINC")
}
