% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_concept.R
\name{download_concept}
\alias{download_concept}
\title{download omop concept file(s) from provided location}
\usage{
download_concept(
  location = "https://omopes.blob.core.windows.net/newcontainer/"
)
}
\arguments{
\item{location}{url of file location, defaults to UCLH filestore}
}
\description{
initially using UCLH temporary private filestore
}
