% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_concepts.R
\name{filter_concepts}
\alias{filter_concepts}
\title{filter omop concepts by standard attributes}
\usage{
filter_concepts(
  df,
  c_ids = NULL,
  d_ids = NULL,
  v_ids = NULL,
  cc_ids = NULL,
  standard = NULL
)
}
\arguments{
\item{df}{dataframe with standard omop concept table columns}

\item{c_ids}{one or more concept_id to filter by, default NULL for all}

\item{d_ids}{one or more domain_id to filter by, default NULL for all}

\item{v_ids}{one or more vocabulary_id to filter by, default NULL for all}

\item{cc_ids}{one or more concept_class_id to filter by, default NULL for all}

\item{standard}{one or more standard_concept to filter by, default NULL for all, S,C}
}
\value{
a filtered dataframe of concepts and attributes
}
\description{
one or more of : concept_id, domain_id, vocabulary_id, concept_class_id, standard_concept
used by other functions
}
\examples{
concept |> filter_concepts(d_ids=c("measurement","drug"),v_ids="SNOMED") |> count(domain_id,vocabulary_id)
}
