% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot_dims.R
\name{get_plot_dims}
\alias{get_plot_dims}
\title{set plot dimensions from a string e.g. "A4" "A4l" etc.}
\usage{
get_plot_dims(canvas)
}
\arguments{
\item{canvas}{string defining canvas size "A4" "A4l" for landscape etc.}
}
\value{
list of width, height, units
}
\description{
options "A4","A4l","A3","A3l","A2","A2l","A1","A1l","A0","A0l","slide","slidehalf"
}
\examples{
get_plot_dims("A1")
}
