% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_omop_name.R
\name{join_omop_name}
\alias{join_omop_name}
\title{join omop concept name onto a dataframe with a *_concept_id column}
\usage{
join_omop_name(df, namestart = NULL)
}
\arguments{
\item{df}{dataframe}

\item{namestart}{start of *_concept_id column, if NULL will join on concept_name}
}
\description{
adds (namestart)_name based on (namestart)_concept_id
e.g. drug_concept_id to get drug_name
}
\examples{
df1 <- tibble(concept_id=(c(3571338L,4002075L)))
join_omop_name(df1)
df2 <- tibble(drug_concept_id=(c(4000794L,4002592L)))
join_omop_name(df2, namestart="drug")
#df2 <- drug_exposure \%>\% distinct(route_concept_id) \%>\% join_omop_name(route_concept_id)
}
