% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_ancestors.R, R/omop_graph.R
\name{omop_ancestors}
\alias{omop_ancestors}
\alias{oance}
\title{find omop concept ancestors of one passed}
\usage{
omop_ancestors(
  c_id = NULL,
  c_ids = NULL,
  d_ids = NULL,
  v_ids = NULL,
  cc_ids = NULL,
  standard = NULL,
  separation = NULL,
  itself = FALSE,
  messages = TRUE
)

oance(
  c_id = NULL,
  c_ids = NULL,
  d_ids = NULL,
  v_ids = NULL,
  cc_ids = NULL,
  standard = NULL,
  separation = NULL,
  itself = FALSE,
  messages = TRUE
)

oance(
  c_id = NULL,
  c_ids = NULL,
  d_ids = NULL,
  v_ids = NULL,
  cc_ids = NULL,
  standard = NULL,
  separation = NULL,
  itself = FALSE,
  messages = TRUE
)
}
\arguments{
\item{c_id}{single omop concept_id or exact concept_name to get ancestors of, default NULL returns all}

\item{c_ids}{one or more concept_id to filter by, default NULL for all}

\item{d_ids}{one or more domain_id to filter by, default NULL for all}

\item{v_ids}{one or more vocabulary_id to filter by, default NULL for all}

\item{cc_ids}{one or more concept_class_id to filter by, default NULL for all}

\item{standard}{one or more standard_concept to filter by, default NULL for all, S,C}

\item{separation}{levels of separation to filter by, default NULL for all}

\item{itself}{whether to include passed concept in returned table (min_levels_of_separation==0), default=FALSE}

\item{messages}{whether to print info messages, default=TRUE}
}
\value{
a dataframe of concepts and attributes
}
\description{
find omop concept ancestors of one passed

super short name func to find ancestors

super short name func to find ancestors
}
\examples{
omop_ancestors(1633308)
#omop_ancestors("Non-invasive blood pressure")
#omop_ancestors("Non-invasive blood pressure",separation=c(1,2))
#epoch_ance <- omop_ancestors("EPOCH, dose-escalated")
#no filtering by ancestors
#cman <- omop_ancestors(v_ids="Cancer Modifier")
# because of R argument matching, you can just use the first unique letters of
# arguments e.g. v for v_ids, cc for cc_ids
chemodrugs <- odesc("Cytotoxic chemotherapeutic", v="HemOnc", d="Regimen")
# because of R argument matching, you can just use the first unique letters of
# arguments e.g. v for v_ids, cc for cc_ids
chemodrugs <- odesc("Cytotoxic chemotherapeutic", v="HemOnc", d="Regimen")
}
