% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_cdm.R
\name{omop_cdm}
\alias{omop_cdm}
\title{read all omop tables from a folder containing parquet or csv files into a list
advantage that you don't need to specify table names}
\usage{
omop_cdm(path, filetype = "parquet", lowercasenames = TRUE)
}
\arguments{
\item{path}{path of folder containing files}

\item{filetype}{default "parquet" option "csv"}

\item{lowercasenames}{whether to make table names lowercase, default TRUE}
}
\value{
a list containing omop tables
}
\description{
read all omop tables from a folder containing parquet or csv files into a list
advantage that you don't need to specify table names
}
