% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_cdm_combine.R
\name{omop_cdm_combine}
\alias{omop_cdm_combine}
\title{combine two omop cdm (Common Data Model) instances
initially works on lists but TODO get it to work on folder names by calling omop_cdm_read()}
\usage{
omop_cdm_combine(cdm1, cdm2, make_person_id_unique = TRUE)
}
\arguments{
\item{cdm1}{list containing cdm tables}

\item{cdm2}{list containing cdm tables}

\item{make_person_id_unique}{whether to check & shift person ids of cdm2, default TRUE}
}
\value{
a list containing merged omop tables
}
\description{
combine two omop cdm (Common Data Model) instances
initially works on lists but TODO get it to work on folder names by calling omop_cdm_read()
}
\examples{
pids1 <- 1:3
pids2 <- 11:12
cdm1 <- list(person=data.frame(person_id=pids1,age=21:23),
             measurement=data.frame(person_id=pids1,meas=5:7))
cdm2 <- list(person=data.frame(person_id=pids2,age=91:92),
            measurement=data.frame(person_id=pids2,meas=15:16),
            death=data.frame(person_id=pids2,d=c(0,1)))
cdm3 <- omop_cdm_combine(cdm1, cdm2)
#test creation of new ids when overlap
pids21 <- 1:2
cdm21 <- list(person=data.frame(person_id=pids21,age=91:92),
            measurement=data.frame(person_id=pids21,meas=15:16),
            death=data.frame(person_id=pids21,d=c(0,1)))
cdm31 <- omop_cdm_combine(cdm1, cdm21)

}
