% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_cdm_merge.R
\name{omop_cdm_merge}
\alias{omop_cdm_merge}
\title{merge two omop cdm (Common Data Model) instances
initially works on lists but TODO get it to work on folder names by calling omop_cdm()}
\usage{
omop_cdm_merge(cdm1, cdm2)
}
\arguments{
\item{cdm1}{list containing cdm tables}

\item{cdm2}{list containing cdm tables}
}
\value{
a list containing merged omop tables
}
\description{
merge two omop cdm (Common Data Model) instances
initially works on lists but TODO get it to work on folder names by calling omop_cdm()
}
