% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_check_names.R
\name{omop_check_names}
\alias{omop_check_names}
\alias{ochecknames}
\title{check that concept names and ids match in a passed table}
\usage{
omop_check_names(
  df,
  id_col_name = "concept_id",
  name_col_name = "concept_name"
)

ochecknames(df, id_col_name = "concept_id", name_col_name = "concept_name")
}
\arguments{
\item{df}{dataframe}

\item{id_col_name}{name of id column, default="concept_id"}

\item{name_col_name}{name of name column, default="concept_name"}
}
\description{
check that concept names and ids match in a passed table

super short name func to check concept names
}
\examples{
#example1 should pass, because names are joined from vocab
data.frame(concept_id=c(3571338L,4002075L)) |> omop_join_name() |> omop_check_names()
#example2 with an incorrect name
tst <- data.frame(concept_id=c(4052465L,4052465L),concept_name=c("test wrong name","Ex-pipe smoker")) |> omop_check_names()
}
