% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_codes.R
\name{omop_codes}
\alias{omop_codes}
\alias{ocodes}
\title{string search of concept_code in omop concepts table}
\usage{
omop_codes(
  findstring,
  ignore_case = TRUE,
  exact = TRUE,
  fixed = FALSE,
  c_ids = NULL,
  d_ids = NULL,
  v_ids = NULL,
  cc_ids = NULL,
  standard = NULL,
  messages = TRUE
)

ocodes(
  findstring,
  ignore_case = TRUE,
  exact = TRUE,
  fixed = FALSE,
  c_ids = NULL,
  d_ids = NULL,
  v_ids = NULL,
  cc_ids = NULL,
  standard = NULL,
  messages = TRUE
)
}
\arguments{
\item{findstring}{string to search for or regex}

\item{ignore_case}{ignore case in string comparison, default TRUE}

\item{exact}{TRUE for exact string search, "start" for exact start, "end" for exact end, default=TRUE because user likely to look for whole code}

\item{fixed}{default FALSE allows regex,TRUE uses grepl exact matching}

\item{c_ids}{one or more concept_id to filter by, default NULL for all}

\item{d_ids}{one or more domain_id to filter by, default NULL for all}

\item{v_ids}{one or more vocabulary_id to filter by, default NULL for all}

\item{cc_ids}{one or more concept_class_id to filter by, default NULL for all}

\item{standard}{one or more standard_concept to filter by, default NULL for all, S,C}

\item{messages}{whether to print info messages, default=TRUE}
}
\description{
string search of concept_code in omop concepts table

super short name func to search concepts by concept_code
}
\examples{
omop_codes("81752-8")
omop_codes("AJCC/UICC-Stage",exact=FALSE)
ocodes("AJCC/UICC-Stage")
}
