% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_concept.R
\name{omop_concept}
\alias{omop_concept}
\title{open a reference to omop concept file(s) from provided location}
\usage{
omop_concept(location = tools::R_user_dir("omopcept", which = "cache"))
}
\arguments{
\item{location}{url of file location, defaults to package cache}
}
\description{
defaults to package cache used by omop_download()
}
\examples{
# open reference, query and collect data to dataframe
omop_concept() |> head() |> dplyr::collect()

}
