% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_concept_ancestor.R
\name{omop_concept_ancestor}
\alias{omop_concept_ancestor}
\alias{oca}
\title{open a reference to omop concept ancestor file from provided location}
\usage{
omop_concept_ancestor(
  location = tools::R_user_dir("omopcept", which = "cache")
)

oca(location = tools::R_user_dir("omopcept", which = "cache"))
}
\arguments{
\item{location}{url of file location, defaults to package cache}
}
\description{
defaults to package cache used by omop_vocab_table_save()
}
\examples{
# open reference, query and collect data to dataframe
omop_concept_ancestor() |> head() |> dplyr::collect()

oca() |> head() |> dplyr::collect()
}
