% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_concept_ancestor_fields.R
\name{omop_concept_ancestor_fields}
\alias{omop_concept_ancestor_fields}
\alias{ocafields}
\title{field names of omop concept ancestor table}
\usage{
omop_concept_ancestor_fields()

ocafields()
}
\description{
field names of omop concept ancestor table

super short name func giving field names of omop concept ancestor table
}
\examples{
omop_concept_ancestor_fields()
}
