% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_concept_fields.R
\name{omop_concept_fields}
\alias{omop_concept_fields}
\alias{ocfields}
\title{field names of omop concept table}
\usage{
omop_concept_fields()

ocfields()
}
\description{
field names of omop concept table

super short name func giving field names of omop concept table
}
\examples{
omop_concept_fields()
}
