% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_concept_relationship.R
\name{omop_concept_relationship}
\alias{omop_concept_relationship}
\alias{ocr}
\title{open a reference to omop concept relationship file from provided location}
\usage{
omop_concept_relationship(
  location = tools::R_user_dir("omopcept", which = "cache")
)

ocr(location = tools::R_user_dir("omopcept", which = "cache"))
}
\arguments{
\item{location}{url of file location, defaults to package cache}
}
\description{
defaults to package cache used by omop_vocab_table_save()
}
\examples{
# open reference, query and collect data to dataframe
omop_concept_relationship() |> head() |> dplyr::collect()

ocr() |> head() |> dplyr::collect()
}
