% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_concept_relationship_fields.R
\name{omop_concept_relationship_fields}
\alias{omop_concept_relationship_fields}
\alias{ocrfields}
\title{field names of omop concept relationship table}
\usage{
omop_concept_relationship_fields()

ocrfields()
}
\description{
field names of omop concept relationship table

super short name func giving field names of omop concept relationship table
}
\examples{
omop_concept_relationship_fields()
}
