% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_domain.R
\name{omop_domain}
\alias{omop_domain}
\title{return domain_id s for concept_id s}
\usage{
omop_domain(c_id)
}
\arguments{
\item{c_id}{concept_id s to get domain_id s for}
}
\description{
return domain_id s for concept_id s
}
\examples{
omop_domain(c(196523,43807321))
#can specify as an integer or not

}
