% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_download.R
\name{omop_download}
\alias{omop_download}
\title{download omop concept file(s) from provided location}
\usage{
omop_download(location = "https://omopes.blob.core.windows.net/newcontainer/")
}
\arguments{
\item{location}{url of file location, defaults to UCLH filestore}
}
\description{
initially using UCLH temporary private filestore
}
