% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_filter_concepts.R
\name{omop_filter_concepts}
\alias{omop_filter_concepts}
\title{filter omop concepts by standard attributes}
\usage{
omop_filter_concepts(
  df,
  c_ids = NULL,
  d_ids = NULL,
  v_ids = NULL,
  cc_ids = NULL,
  standard = NULL
)
}
\arguments{
\item{df}{dataframe with standard omop concept table columns}

\item{c_ids}{one or more concept_id to filter by, default NULL for all}

\item{d_ids}{one or more domain_id to filter by, default NULL for all}

\item{v_ids}{one or more vocabulary_id to filter by, default NULL for all}

\item{cc_ids}{one or more concept_class_id to filter by, default NULL for all}

\item{standard}{one or more standard_concept to filter by, default NULL for all, S,C}
}
\value{
a filtered dataframe of concepts and attributes
}
\description{
Works on a dataframe or an arrow object. Can filter by one or more of : concept_id, domain_id, vocabulary_id, concept_class_id, standard_concept
used by other functions
}
\examples{
open_concept() |>
  omop_filter_concepts(d_ids=c("measurement","drug"),v_ids="SNOMED") |>
  dplyr::collect() |>
  dplyr::count(domain_id,vocabulary_id)
open_concept() |> omop_filter_concepts(v_ids="Gender") |> dplyr::collect()
}
