% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_names.R
\name{omop_names}
\alias{omop_names}
\alias{onames}
\title{string search of concept_name in omop concepts table}
\usage{
omop_names(
  findstring,
  ignore_case = TRUE,
  c_ids = NULL,
  d_ids = NULL,
  v_ids = NULL,
  cc_ids = NULL,
  standard = NULL
)

onames(
  findstring,
  ignore_case = TRUE,
  c_ids = NULL,
  d_ids = NULL,
  v_ids = NULL,
  cc_ids = NULL,
  standard = NULL
)
}
\arguments{
\item{findstring}{to search for or regex}

\item{ignore_case}{ignore case in string comparison, default TRUE}

\item{c_ids}{one or more concept_id to filter by, default NULL for all}

\item{d_ids}{one or more domain_id to filter by, default NULL for all}

\item{v_ids}{one or more vocabulary_id to filter by, default NULL for all}

\item{cc_ids}{one or more concept_class_id to filter by, default NULL for all}

\item{standard}{one or more standard_concept to filter by, default NULL for all, S,C}
}
\description{
string search of concept_name in omop concepts table

super short name func to search concepts by concept_name
}
\examples{
omop_names("AJCC/UICC Stage")
omop_names("chemotherapy", v_ids="LOINC")
omop_names("chemotherapy", v_ids=c("LOINC","SNOMED"), d_ids=c("Observation","Procedure"))
onames("chemotherapy", v_ids="LOINC")
}
