% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_relations1step.R
\name{omop_relations1step}
\alias{omop_relations1step}
\title{find immediate omop concept relations of one passed
ADVISORY instead to use omop_relations() with default nsteps=1 arg.}
\usage{
omop_relations1step(
  c_id = NULL,
  c_ids = NULL,
  d_ids = NULL,
  v_ids = NULL,
  cc_ids = NULL,
  standard = NULL,
  r_ids = NULL,
  itself = FALSE,
  names2avoid = c("SNOMED CT core", "Defined", "Primitive"),
  join_names = TRUE,
  messages = TRUE
)
}
\arguments{
\item{c_id}{single omop concept_id or exact concept_name to get relations of, default NULL returns all}

\item{c_ids}{one or more concept_id to filter by, default NULL for all}

\item{d_ids}{one or more domain_id to filter by, default NULL for all}

\item{v_ids}{one or more vocabulary_id to filter by, default NULL for all}

\item{cc_ids}{one or more concept_class_id to filter by, default NULL for all}

\item{standard}{one or more standard_concept to filter by, default NULL for all, S,C}

\item{r_ids}{one or more relationship_id to filter by, default NULL for all, e.g c('Is a','Subsumes')}

\item{itself}{whether to include relations to concept itself, default=FALSE}

\item{names2avoid}{concept names to avoid, defaults to generic concepts with lots relations, can be set to NULL}

\item{join_names}{whether to join concept names onto ids, default TRUE, FALSE used by omop_relations() to speed up}

\item{messages}{whether to print info messages, default=TRUE}
}
\value{
a dataframe of concepts and attributes
}
\description{
find immediate omop concept relations of one passed
ADVISORY instead to use omop_relations() with default nsteps=1 arg.
}
\examples{
omop_relations1step("Non-invasive blood pressure")
}
