% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_relations_multiple.R
\name{omop_relations_multiple}
\alias{omop_relations_multiple}
\title{extract omop concept relations of a vector of concept_ids - immediate relations with indication of relationship}
\usage{
omop_relations_multiple(
  mc_ids,
  c_ids = NULL,
  d_ids = NULL,
  v_ids = NULL,
  cc_ids = NULL,
  standard = NULL,
  r_ids = NULL,
  messages = TRUE,
  num_recurse = 1
)
}
\arguments{
\item{mc_ids}{a vector of multiple concept_id's to get relations of}

\item{c_ids}{one or more concept_id to filter by, default NULL for all}

\item{d_ids}{one or more domain_id to filter by, default NULL for all}

\item{v_ids}{one or more vocabulary_id to filter by, default NULL for all}

\item{cc_ids}{one or more concept_class_id to filter by, default NULL for all}

\item{standard}{one or more standard_concept to filter by, default NULL for all, S,C}

\item{r_ids}{one or more relationship_id to filter by, default NULL for all, e.g c('Is a','Subsumes')}

\item{messages}{whether to print info messages, default=TRUE}

\item{num_recurse}{number of recursions to search}
}
\value{
a dataframe of concepts and attributes
}
\description{
extract omop concept relations of a vector of concept_ids - immediate relations with indication of relationship
}
\examples{
orm <- omop_relations_multiple(c(3571338L,3655355L), r_ids=c('Is a','Subsumes'), num_recurse=1)
#omop_relations_multiple(c(3571338L,3655355L), r_ids=c('Is a','Subsumes'), num_recurse=2)
}
