% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_relations_names.R
\name{omop_relations_names}
\alias{omop_relations_names}
\alias{ornames}
\title{string search of concepts by name in the CONCEPT_RELATIONSHIP table}
\usage{
omop_relations_names(
  findstring,
  ignore_case = TRUE,
  exact = FALSE,
  fixed = FALSE,
  c_ids = NULL,
  d_ids = NULL,
  v_ids = NULL,
  cc_ids = NULL,
  standard = NULL,
  messages = TRUE
)

ornames(
  findstring,
  ignore_case = TRUE,
  exact = FALSE,
  fixed = FALSE,
  c_ids = NULL,
  d_ids = NULL,
  v_ids = NULL,
  cc_ids = NULL,
  standard = NULL,
  messages = TRUE
)
}
\arguments{
\item{findstring}{to search for or regex, e.g. "^a" to find those starting with A}

\item{ignore_case}{ignore case in string comparison, default TRUE}

\item{exact}{TRUE for exact string search, "start" for exact start, "end" for exact end, default=FALSE for str_detect}

\item{fixed}{default FALSE allows regex,TRUE uses grepl exact matching}

\item{c_ids}{one or more concept_id to filter by, default NULL for all}

\item{d_ids}{one or more domain_id to filter by, default NULL for all}

\item{v_ids}{one or more vocabulary_id to filter by, default NULL for all}

\item{cc_ids}{one or more concept_class_id to filter by, default NULL for all}

\item{standard}{one or more standard_concept to filter by, default NULL for all, S,C}

\item{messages}{whether to print info messages, default=TRUE}
}
\description{
a different (and quicker) way of finding relationships than omop_relations() that gets relations of a specified concept
note that names are not contained in CONCEPT_RELATIONSHIP so the function joins to CONCEPT to query
}
\examples{
orn <- omop_relations_names("diffuse large B cell lymphoma")
#omop_relations_names("AJCC/UICC Stage")
#omop_relations_names("chemotherapy", v_ids="LOINC")
#omop_relations_names("chemotherapy", v_ids=c("LOINC","SNOMED"), d_ids=c("Observation","Procedure"))
#set the findstring to "" to get all rows satisfying the other conditions
#omop_relations_names("", v_ids="Gender")
#omop_relations_names("", d_ids="Type Concept", standard="S")
}
