% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_relations_recursive.R
\name{omop_relations_recursive}
\alias{omop_relations_recursive}
\title{recursively extract omop concept relations of one passed - immediate relations with indication of relationship}
\usage{
omop_relations_recursive(
  c_id = NULL,
  c_ids = NULL,
  d_ids = NULL,
  v_ids = NULL,
  cc_ids = NULL,
  standard = NULL,
  r_ids = NULL,
  itself = FALSE,
  names2avoid = c("SNOMED CT core", "Defined", "Primitive"),
  messages = TRUE,
  nsteps = 1,
  add_step_column = TRUE
)
}
\arguments{
\item{c_id}{single omop concept_id or exact concept_name to get relations of, default NULL returns all}

\item{c_ids}{one or more concept_id to filter by, default NULL for all}

\item{d_ids}{one or more domain_id to filter by, default NULL for all}

\item{v_ids}{one or more vocabulary_id to filter by, default NULL for all}

\item{cc_ids}{one or more concept_class_id to filter by, default NULL for all}

\item{standard}{one or more standard_concept to filter by, default NULL for all, S,C}

\item{r_ids}{one or more relationship_id to filter by, default NULL for all, e.g c('Is a','Subsumes')}

\item{itself}{whether to include relations to concept itself, default=FALSE}

\item{names2avoid}{concept names to avoid, defaults to generic concepts with lots relations, can be set to NULL}

\item{messages}{whether to print info messages, default=TRUE}

\item{nsteps}{number of recursions to search}

\item{add_step_column}{whether to add column with step level, default=TRUE}
}
\value{
a dataframe of concepts and attributes
}
\description{
recursively extract omop concept relations of one passed - immediate relations with indication of relationship
}
