% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_table.R
\name{omop_table}
\alias{omop_table}
\title{read single omop table from parquet or csv file}
\usage{
omop_table(tablename, path, filetype = "parquet")
}
\arguments{
\item{tablename}{name of omop table e.g. "person"}

\item{path}{file path}

\item{filetype}{default "parquet" option "csv"}
}
\value{
a dataframe
}
\description{
read single omop table from parquet or csv file
}
