% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_vocab_table_save.R
\name{omop_vocab_table_save}
\alias{omop_vocab_table_save}
\title{save omop vocabulary parquet file from provided location to local package cache}
\usage{
omop_vocab_table_save(
  tablename = "concept",
  from = NULL,
  to = tools::R_user_dir("omopcept", which = "cache"),
  download_max_secs = 720
)
}
\arguments{
\item{tablename}{which omop table to download, defaults to 'concept'}

\item{from}{url of file source location, defaults to UCLH filestore}

\item{to}{path to save file locally, defaults to package cache which is where omopcept functions look}

\item{download_max_secs}{max download secounds, default 720 = 12 mins}
}
\value{
nothing
}
\description{
default using UCLH temporary private filestore, option to use local files
}
\seealso{
\code{\link[=omop_vocabs_preprocess]{omop_vocabs_preprocess()}}
alternative that reads in csvs downloaded from Athena and saves to parquet also in local package cache
}
