% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_vocabs_preprocess.R
\name{omop_vocabs_preprocess}
\alias{omop_vocabs_preprocess}
\title{read in omop vocabulary csvs downloaded from Athena, preprocess
and write out as binary parquet files for better performance.}
\usage{
omop_vocabs_preprocess(athena_source_directory)
}
\arguments{
\item{athena_source_directory}{filepath where omop vocab csvs are stored}
}
\description{
TODO think: do I want to save parquets directly into the package cache from here
}
