% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omopcept-package.R
\docType{package}
\name{omopcept-package}
\alias{omopcept}
\alias{omopcept-package}
\title{omopcept: Tools to query OMOP conCEPTs for standard health vocabularies (pros without the cons!)}
\description{
Access to OMOP Common Data Model medical vocabularies and flexible tidyverse compatible R functions for querying.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andysouth/omopcept}
  \item Report bugs at \url{https://github.com/andysouth/omopcept/issues}
}

}
\author{
\strong{Maintainer}: Andy South \email{southandy@gmail.com} (\href{https://orcid.org/0000-0003-4051-6135}{ORCID})

}
\keyword{internal}
