% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omopcepts-package.R
\docType{package}
\name{omopcepts-package}
\alias{omopcepts}
\alias{omopcepts-package}
\title{omopcepts: Tools to query OMOP medical vocabularies}
\description{
Access to a subset of OMOP Common Data Model medical vocabularies and flexible tidyverse compatible R functions for querying.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andysouth/omopcepts}
  \item Report bugs at \url{https://github.com/andysouth/omopcepts/issues}
}

}
\author{
\strong{Maintainer}: Andy South \email{southandy@gmail.com} (\href{https://orcid.org/0000-0003-4051-6135}{ORCID})

}
\keyword{internal}
