% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omopfreq.R
\name{omopfreqconceptclass}
\alias{omopfreqconceptclass}
\alias{ofreqcc}
\title{count freq of values in an omop field
EXPERIMENTAL may change}
\usage{
omopfreqconceptclass(df1, sort = TRUE)

ofreqcc(df1, sort = TRUE)
}
\arguments{
\item{df1}{an omop table}

\item{sort}{whether to sort freqs, default TRUE}
}
\description{
count freq of values in an omop field
EXPERIMENTAL may change

super short name func to count freq of values in an omop field
}
\examples{
omop_relations("Non-invasive blood pressure") |> omopfreqconceptclass()

}
