% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omopfreq.R
\name{omopfreqrelationship}
\alias{omopfreqrelationship}
\alias{ofreqr}
\title{count freq of values in an omop field
EXPERIMENTAL may change}
\usage{
omopfreqrelationship(df1, sort = TRUE)

ofreqr(df1, sort = TRUE)
}
\arguments{
\item{df1}{an omop table}

\item{sort}{whether to sort freqs, default TRUE}
}
\description{
count freq of values in an omop field
EXPERIMENTAL may change

super short name func to count freq of values in an omop field
}
\examples{
omop_relations("Non-invasive blood pressure") |> omopfreqrelationship()

}
