% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_concept.R
\name{open_concept}
\alias{open_concept}
\title{open a reference to omop concept file(s) from provided location}
\usage{
open_concept(location = tools::R_user_dir("omopcepts", which = "cache"))
}
\arguments{
\item{location}{url of file location, defaults to package cache}
}
\description{
initially using R user cache
}
\examples{
# open refernce, query and collect data to dataframe
open_concept() |> head() |> dplyr::collect()

}
