% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pipe_if}
\alias{pipe_if}
\title{conditionally apply a pipe}
\usage{
pipe_if(df, cond, func)
}
\arguments{
\item{df}{dataframe to apply funct to if cond TRUE}

\item{cond}{condition whether to run func, not vectorised (should be a single logical)}

\item{func}{function to run}
}
\description{
conditionally apply a pipe
}
\examples{
data.frame(a=c(1:5)) |> pipe_if(TRUE, \(d) d |> mutate(b=a*2))
}
