% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_c_id.R
\name{check_c_id}
\alias{check_c_id}
\title{check c_id arg to omop_ancestors() and omop_descendants()
TODO want to not export but couldn't get to work when not exported}
\usage{
check_c_id(c_id)
}
\arguments{
\item{c_id}{arg to check : single omop concept_id or exact concept_name to get ancestors of}
}
\description{
check c_id arg to omop_ancestors() and omop_descendants()
TODO want to not export but couldn't get to work when not exported
}
\examples{
# example code
c_id <- 1633308
check_c_id(c_id)

}
