% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dynamic_by}
\alias{dynamic_by}
\title{Produces a list from string variables suitable for use in join expressions
For example, (by=dynamic_by(variable,"rhs"))}
\usage{
dynamic_by(lhs, rhs)
}
\description{
Produces a list from string variables suitable for use in join expressions
For example, (by=dynamic_by(variable,"rhs"))
}
