% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_omop_name.R
\name{join_omop_name}
\alias{join_omop_name}
\title{join omop concept name onto a dataframe with a *_concept_id column}
\usage{
join_omop_name(df, namestart = NULL)
}
\arguments{
\item{df}{dataframe}

\item{namestart}{start of *_concept_id column, if NULL will join on concept_name}
}
\description{
joins on \link{arg}_name based on \link{arg}_concept_id
e.g. drug_concept_id to get drug_name
}
