% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_ancestors.R
\name{omop_ancestors}
\alias{omop_ancestors}
\title{find omop concept ancestors of one passed}
\usage{
omop_ancestors(
  c_id,
  c_ids = NULL,
  d_ids = NULL,
  v_ids = NULL,
  cc_ids = NULL,
  standard = NULL
)
}
\arguments{
\item{c_id}{single omop concept_id or exact concept_name to get ancestors of}

\item{c_ids}{one or more concept_id to filter by, default NULL for all}

\item{d_ids}{one or more domain_id to filter by, default NULL for all}

\item{v_ids}{one or more vocabulary_id to filter by, default NULL for all}

\item{cc_ids}{one or more concept_class_id to filter by, default NULL for all}

\item{standard}{one or more standard_concept to filter by, default NULL for all, S,C}
}
\value{
a dataframe of concepts and attributes
}
\description{
find omop concept ancestors of one passed
}
\examples{
omop_ancestors(1633308)
}
