% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_ancestors.R
\name{omop_ancestors}
\alias{omop_ancestors}
\alias{oance}
\title{find omop concept ancestors of one passed}
\usage{
omop_ancestors(
  c_id,
  c_ids = NULL,
  d_ids = NULL,
  v_ids = NULL,
  cc_ids = NULL,
  standard = NULL,
  messages = TRUE
)

oance(
  c_id,
  c_ids = NULL,
  d_ids = NULL,
  v_ids = NULL,
  cc_ids = NULL,
  standard = NULL,
  messages = TRUE
)
}
\arguments{
\item{c_id}{single omop concept_id or exact concept_name to get ancestors of}

\item{c_ids}{one or more concept_id to filter by, default NULL for all}

\item{d_ids}{one or more domain_id to filter by, default NULL for all}

\item{v_ids}{one or more vocabulary_id to filter by, default NULL for all}

\item{cc_ids}{one or more concept_class_id to filter by, default NULL for all}

\item{standard}{one or more standard_concept to filter by, default NULL for all, S,C}

\item{messages}{whether to print info messages, default=TRUE}
}
\value{
a dataframe of concepts and attributes
}
\description{
find omop concept ancestors of one passed

super short name func to find ancestors
}
\examples{
omop_ancestors(1633308)
omop_ancestors("Non-invasive blood pressure")
epoch_ance <- omop_ancestors("EPOCH, dose-escalated")
# because of R argument matching, you can just use the first unique letters of
# arguments e.g. v for v_ids, cc for cc_ids
chemodrugs <- odesc("Cytotoxic chemotherapeutic", v="HemOnc", d="Regimen")
}
