% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_cdm_combine.R
\name{omop_cdm_combine}
\alias{omop_cdm_combine}
\title{combine two omop cdm (Common Data Model) instances
initially works on lists
NOTE these omop_cdm* functions are for omop extracts rather than the concepts and may be best moved to another package
TODO could be modified relatively easily to work on >2 extracts up to 9 :-)
TODO get it to work on folder names by calling omop_cdm_read()}
\usage{
omop_cdm_combine(
  cdm1,
  cdm2,
  make_care_site_id_unique = TRUE,
  make_ids_unique = TRUE,
  add_care_site_name_to_person_id_tables = TRUE
)
}
\arguments{
\item{cdm1}{list containing cdm tables}

\item{cdm2}{list containing cdm tables}

\item{make_care_site_id_unique}{whether to set care_site_id to 1 & 2 for each instance, default TRUE}

\item{make_ids_unique}{whether to uniqueify all other IDs (multiply by 10 & add extractnum), default TRUE}

\item{add_care_site_name_to_person_id_tables}{whether to add column to all tables with person_id}
}
\value{
a list containing merged omop tables
}
\description{
combine two omop cdm (Common Data Model) instances
initially works on lists
NOTE these omop_cdm* functions are for omop extracts rather than the concepts and may be best moved to another package
TODO could be modified relatively easily to work on >2 extracts up to 9 :-)
TODO get it to work on folder names by calling omop_cdm_read()
}
\examples{
#test creation of new ids when overlap
pids1 <- 1:3
pids2 <- 1:2
cdm1 <- list(person=data.frame(person_id=pids1,age=21:23,care_site_id=1L),
             measurement=data.frame(person_id=pids1,meas=5:7),
             care_site=data.frame(care_site_id=1L, care_site_name="hospital1"))
cdm2 <- list(person=data.frame(person_id=pids2,age=91:92,care_site_id=2L),
            measurement=data.frame(person_id=pids2,meas=15:16),
            care_site=data.frame(care_site_id=1L, care_site_name="hospital2"),
            death=data.frame(person_id=pids2,d=c(0,1)))
cdm3 <- omop_cdm_combine(cdm1, cdm2)

}
