% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_cdm_combine.R
\name{omop_cdm_combine}
\alias{omop_cdm_combine}
\title{combine two omop cdm (Common Data Model) instances
initially works on lists but TODO get it to work on folder names by calling omop_cdm_read()}
\usage{
omop_cdm_combine(cdm1, cdm2)
}
\arguments{
\item{cdm1}{list containing cdm tables}

\item{cdm2}{list containing cdm tables}
}
\value{
a list containing merged omop tables
}
\description{
combine two omop cdm (Common Data Model) instances
initially works on lists but TODO get it to work on folder names by calling omop_cdm_read()
}
\examples{
cdm1 <- list(person=tibble(id=1:3,age=21:23),
             measurement=tibble(id=1:3,meas=5:7))
cdm2 <- list(person=tibble(id=11:12,age=91:92),
            measurement=tibble(id=11:12,meas=15:16),
            death=tibble(id=11:12,d=c(0,1)))
cdm3 <- omop_cdm_combine(cdm1, cdm2)

}
