% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_descendants.R
\name{omop_descendants}
\alias{omop_descendants}
\alias{odesc}
\title{find omop concept descendants of one passed}
\usage{
omop_descendants(
  c_id,
  c_ids = NULL,
  d_ids = NULL,
  v_ids = NULL,
  cc_ids = NULL,
  standard = NULL,
  itself = FALSE,
  messages = TRUE
)

odesc(
  c_id,
  c_ids = NULL,
  d_ids = NULL,
  v_ids = NULL,
  cc_ids = NULL,
  standard = NULL,
  itself = FALSE,
  messages = TRUE
)
}
\arguments{
\item{c_id}{single omop concept_id or exact concept_name to get descendants of}

\item{c_ids}{one or more concept_id to filter by, default NULL for all}

\item{d_ids}{one or more domain_id to filter by, default NULL for all}

\item{v_ids}{one or more vocabulary_id to filter by, default NULL for all}

\item{cc_ids}{one or more concept_class_id to filter by, default NULL for all}

\item{standard}{one or more standard_concept to filter by, default NULL for all, S,C}

\item{itself}{whether to include passed concept in returned table (min_levels_of_separation==0), default=FALSE}

\item{messages}{whether to print info messages, default=TRUE}
}
\value{
a dataframe of concepts and attributes
}
\description{
find omop concept descendants of one passed

super short name func to find descendants
}
\examples{
omop_descendants(1633308)
#omop_descendants("lenalidomide")
#omop_descendants("Non-invasive blood pressure")
#chemodrugs <- omop_descendants("Cytotoxic chemotherapeutic",v_ids="HemOnc",d_ids="Regimen")
# because of R argument matching, you can just use the first unique letters of
# arguments e.g. v for v_ids, cc for cc_ids
chemodrugs <- odesc("Cytotoxic chemotherapeutic", v="HemOnc", d="Regimen")
}
