% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_graph.R
\name{omop_graph}
\alias{omop_graph}
\title{ggraph omop hierarchy}
\usage{
omop_graph(
  c_id,
  c_ids = NULL,
  d_ids = NULL,
  v_ids = NULL,
  cc_ids = NULL,
  standard = NULL,
  separation = NULL,
  itself = FALSE,
  messages = TRUE
)
}
\arguments{
\item{c_id}{single omop concept_id or exact concept_name to get ancestors of}

\item{c_ids}{one or more concept_id to filter by, default NULL for all}

\item{d_ids}{one or more domain_id to filter by, default NULL for all}

\item{v_ids}{one or more vocabulary_id to filter by, default NULL for all}

\item{cc_ids}{one or more concept_class_id to filter by, default NULL for all}

\item{standard}{one or more standard_concept to filter by, default NULL for all, S,C}

\item{separation}{levels of separation to filter by, default NULL for all}

\item{itself}{whether to include passed concept in returned table (min_levels_of_separation==0), default=FALSE}

\item{messages}{whether to print info messages, default=TRUE}
}
\value{
a dataframe of concepts and attributes
}
\description{
ggraph omop hierarchy
}
\examples{
#TODO enable example when working
#omop_graph(1633308)
#omop_graph("Non-invasive blood pressure")
#omop_graph("Non-invasive blood pressure",separation=c(1,2))
#epoch_ance <- omop_graph("EPOCH, dose-escalated")
}
