% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_graph.R
\name{omop_graph}
\alias{omop_graph}
\title{graph omop hierarchy
accepts output from either omop_ancestors(), omop_descendants() or omop_relations}
\usage{
omop_graph(
  dfin,
  ggrlayout = "graphopt",
  palettebrewer = "Set1",
  edgecolour = "grey71",
  nodecolourvar = "domain_id",
  textcolourvar = NULL,
  nodetxtangle = 0,
  nodetxtsize = 9,
  legendtxtsize = 18,
  titletxtsize = 18,
  legendshow = TRUE,
  legendpos = "bottom",
  legenddir = "horizontal",
  legendcm = 3,
  saveplot = TRUE,
  filetype = "pdf",
  filenameroot = "omop_graph",
  filenamecustom = NULL,
  width = 50,
  height = 30,
  units = "cm",
  titlecolour = "darkred",
  graphtitle = "omopcept graph",
  graphsubtitle = NULL,
  plot = TRUE,
  messages = TRUE
)
}
\arguments{
\item{dfin}{dataframe output from either omop_ancestors(), omop_descendants() or omop_relations}

\item{ggrlayout}{ggraph layout, default = 'graphopt'}

\item{palettebrewer}{colour brewer pallette, default='Set1', other options e.g. 'Dark2' see RColorBrewer::brewer.pal.info}

\item{edgecolour}{colour for lines joining nodes}

\item{nodecolourvar}{column to specify node colour, default="domain_id" other options "vocabulary_id" "concept_class_id" "standard_concept"}

\item{textcolourvar}{column to specify node text colour, default=NULL then set same as node_colour above. Other options "vocabulary_id" "concept_class_id" "standard_concept"}

\item{nodetxtangle}{node text angle, default=0, 90 gives vertical text}

\item{nodetxtsize}{node text size, default=9}

\item{legendtxtsize}{text size for legend, default=20}

\item{titletxtsize}{text size for title, default=20}

\item{legendshow}{whether to show legend, default TRUE}

\item{legendpos}{legend position, default 'bottom'}

\item{legenddir}{legen direction default = 'horizontal'}

\item{legendcm}{legend size cm, default=3}

\item{saveplot}{whether to save plot, default TRUE}

\item{filetype}{output image file, default='pdf'}

\item{filenameroot}{optional root for an auto filename for plot (not used if filenamecustom is supplied)}

\item{filenamecustom}{optional filename for plot, otherwise default name is created}

\item{width}{plot width, default=50}

\item{height}{plot height, default=30}

\item{units}{plot size units default='cm'}

\item{titlecolour}{colour for main title, default='darkred'}

\item{graphtitle}{optional title for graph, default NULL for none}

\item{graphsubtitle}{optional subtitle for graph, default NULL for none}

\item{plot}{whether to display plot, default TRUE, but note that large plots will not display well in R graphics window but do output well to pdf}

\item{messages}{whether to print info messages, default=TRUE}
}
\value{
ggraph object
}
\description{
graph omop hierarchy
accepts output from either omop_ancestors(), omop_descendants() or omop_relations
}
\examples{
#pressure <- omop_names("^Blood pressure$",standard='S')
#press_descend <- omop_descendants(pressure$concept_id[1])
#omop_graph(press_descend, filenameroot="bloodpressure",graphtitle="OMOP Blood Pressure")
}
