% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_graph.R
\name{omop_graph_calc}
\alias{omop_graph_calc}
\title{calculate nodes and edges from omop hierarchy
accepts output from either omop_ancestors(), omop_descendants() or omop_relations
used by omop_graph(), you are only likely to want to use on it's own to
a) separate calculation & visualisation so that you can join attributes for visualisation
b) pass the nodes and edges to a different graph rendering package}
\usage{
omop_graph_calc(dfin)
}
\arguments{
\item{dfin}{dataframe output from either omop_ancestors(), omop_descendants() or omop_relations}
}
\value{
list containing edges & nodes tables
}
\description{
calculate nodes and edges from omop hierarchy
accepts output from either omop_ancestors(), omop_descendants() or omop_relations
used by omop_graph(), you are only likely to want to use on it's own to
a) separate calculation & visualisation so that you can join attributes for visualisation
b) pass the nodes and edges to a different graph rendering package
}
\examples{
bp <- omop_relations("Non-invasive blood pressure")
listedges_nodes <- omop_graph_calc(bp)
rel <- omop_concept_relationship() |> head() |> collect()
listrel <- omop_graph_calc(rel)
}
