% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_id.R
\name{omop_id}
\alias{omop_id}
\title{query a concept record by concept_id from omop concepts table}
\usage{
omop_id(c_id)
}
\arguments{
\item{c_id}{one concept_id to filter by}
}
\description{
query a concept record by concept_id from omop concepts table
}
\examples{
omop_id(3807321L)

}
