% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_join_name.R
\name{omop_join_name}
\alias{omop_join_name}
\alias{ojoin}
\title{join omop concept name onto a dataframe with a *_concept_id column}
\usage{
omop_join_name(
  df,
  namestart = "",
  namefull = "",
  domain = FALSE,
  vocabulary = FALSE,
  concept_class = FALSE,
  concept_code = FALSE
)

ojoin(
  df,
  namestart = "",
  namefull = "",
  domain = FALSE,
  vocabulary = FALSE,
  concept_class = FALSE,
  concept_code = FALSE
)
}
\arguments{
\item{df}{dataframe}

\item{namestart}{start of *_concept_id column, if "" will join on concept_name, ignored if namefull used}

\item{namefull}{optional full name of concept_id column, if default "" namestart used}

\item{domain}{whether to add domain_id column, default FALSE}

\item{vocabulary}{whether to add vocabulary_id column, default FALSE}

\item{concept_class}{whether to add concept_class_id column, default FALSE}

\item{concept_code}{whether to add concept_code column, default FALSE}
}
\description{
adds (namestart)_name based on (namestart)_concept_id
e.g. drug_concept_id to get drug_name
}
\examples{
data.frame(concept_id=(c(3571338L,4002075L))) |> omop_join_name()
data.frame(drug_concept_id=(c(4000794L,4002592L))) |> omop_join_name(namestart="drug")
#if drug_exposure loaded:
#df2 <- drug_exposure |> distinct(route_concept_id) |> omop_join_name(namestart="route")
#df3 <- omop_concept_relationship() |> head() |>
#       dplyr::collect() |> omop_join_name(namefull="concept_id_1")
data.frame(drug_concept_id=(c(4000794L,4002592L))) |> ojoin(namestart="drug")
}
