% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_names.R, R/omop_relations_names.R
\name{omop_names}
\alias{omop_names}
\alias{onames}
\title{string search of concept_name in omop concepts table}
\usage{
omop_names(
  findstring,
  ignore_case = TRUE,
  exact = FALSE,
  fixed = FALSE,
  c_ids = NULL,
  d_ids = NULL,
  v_ids = NULL,
  cc_ids = NULL,
  standard = NULL,
  messages = TRUE
)

onames(
  findstring,
  ignore_case = TRUE,
  exact = FALSE,
  fixed = FALSE,
  c_ids = NULL,
  d_ids = NULL,
  v_ids = NULL,
  cc_ids = NULL,
  standard = NULL,
  messages = TRUE
)

omop_names(
  findstring,
  ignore_case = TRUE,
  exact = FALSE,
  fixed = FALSE,
  c_ids = NULL,
  d_ids = NULL,
  v_ids = NULL,
  cc_ids = NULL,
  standard = NULL,
  messages = TRUE
)

onames(
  findstring,
  ignore_case = TRUE,
  exact = FALSE,
  fixed = FALSE,
  c_ids = NULL,
  d_ids = NULL,
  v_ids = NULL,
  cc_ids = NULL,
  standard = NULL,
  messages = TRUE
)
}
\arguments{
\item{findstring}{to search for or regex, e.g. "^a" to find those starting with A}

\item{ignore_case}{ignore case in string comparison, default TRUE}

\item{exact}{TRUE for exact string search, "start" for exact start, "end" for exact end, default=FALSE for str_detect}

\item{fixed}{default FALSE allows regex,TRUE uses grepl exact matching}

\item{c_ids}{one or more concept_id to filter by, default NULL for all}

\item{d_ids}{one or more domain_id to filter by, default NULL for all}

\item{v_ids}{one or more vocabulary_id to filter by, default NULL for all}

\item{cc_ids}{one or more concept_class_id to filter by, default NULL for all}

\item{standard}{one or more standard_concept to filter by, default NULL for all, S,C}

\item{messages}{whether to print info messages, default=TRUE}
}
\description{
note that names are not contained in CONCEPT_RELATIONSHIP so the function joins to CONCEPT to query
}
\examples{
omop_names("AJCC/UICC Stage")
#omop_names("chemotherapy", v_ids="LOINC")
#omop_names("chemotherapy", v_ids=c("LOINC","SNOMED"), d_ids=c("Observation","Procedure"))
#set the findstring to "" to get all rows satisfying the other conditions
#omop_names("", v_ids="Gender")
#omop_names("", d_ids="Type Concept", standard="S")
#exact= options
#t1 <- onames("tobacco")
#returning 616 concepts
#t2 <- onames("tobacco",exact=TRUE)
#returning 2 concepts
#t3 <- onames("tobacco",exact="start")
#returning 229 concepts
#t4 <- onames("tobacco",exact="end")
#returning 54 concepts
# onames("chemotherapy", v_ids="LOINC")
# because of R argument matching, you can just use the first unique letters of
# arguments e.g. v for v_ids, cc for cc_ids
# to get all clinical drugs starting with A
onames("^a", d="DRUG", v="SNOMED", cc="Clinical Drug")
# to get all 'chop' cancer regimens
#chops <- onames("chop", d="Regimen")
omop_names("AJCC/UICC Stage")
#omop_names("chemotherapy", v_ids="LOINC")
#omop_names("chemotherapy", v_ids=c("LOINC","SNOMED"), d_ids=c("Observation","Procedure"))
#set the findstring to "" to get all rows satisfying the other conditions
#omop_names("", v_ids="Gender")
#omop_names("", d_ids="Type Concept", standard="S")
#exact= options
#t1 <- onames("tobacco")
#returning 616 concepts
#t2 <- onames("tobacco",exact=TRUE)
#returning 2 concepts
#t3 <- onames("tobacco",exact="start")
#returning 229 concepts
#t4 <- onames("tobacco",exact="end")
#returning 54 concepts
# onames("chemotherapy", v_ids="LOINC")
# because of R argument matching, you can just use the first unique letters of
# arguments e.g. v for v_ids, cc for cc_ids
# to get all clinical drugs starting with A
onames("^a", d="DRUG", v="SNOMED", cc="Clinical Drug")
# to get all 'chop' cancer regimens
#chops <- onames("chop", d="Regimen")
}
