% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_relations.R
\name{omop_relations}
\alias{omop_relations}
\alias{orels}
\title{recursively extract omop concept relations of one passed - immediate relations with indication of relationship}
\usage{
omop_relations(
  c_id = NULL,
  c_ids = NULL,
  d_ids = NULL,
  v_ids = NULL,
  cc_ids = NULL,
  standard = NULL,
  r_ids = NULL,
  itself = FALSE,
  names2avoid = c("SNOMED CT core", "Defined", "Primitive"),
  messages = TRUE,
  nsteps = 1,
  add_step_column = TRUE
)

orels(
  c_id = NULL,
  c_ids = NULL,
  d_ids = NULL,
  v_ids = NULL,
  cc_ids = NULL,
  standard = NULL,
  r_ids = NULL,
  itself = FALSE,
  names2avoid = c("SNOMED CT core", "Defined", "Primitive"),
  messages = TRUE,
  nsteps = 1,
  add_step_column = TRUE
)
}
\arguments{
\item{c_id}{single omop concept_id or exact concept_name to get relations of, default NULL returns all}

\item{c_ids}{one or more concept_id to filter by, default NULL for all}

\item{d_ids}{one or more domain_id to filter by, default NULL for all}

\item{v_ids}{one or more vocabulary_id to filter by, default NULL for all}

\item{cc_ids}{one or more concept_class_id to filter by, default NULL for all}

\item{standard}{one or more standard_concept to filter by, default NULL for all, S,C}

\item{r_ids}{one or more relationship_id to filter by, default NULL for all, e.g c('Is a','Subsumes')}

\item{itself}{whether to include relations to concept itself, default=FALSE}

\item{names2avoid}{concept names to avoid, defaults to generic concepts with lots relations, can be set to NULL}

\item{messages}{whether to print info messages, default=TRUE}

\item{nsteps}{number of recursions to search}

\item{add_step_column}{whether to add column with step level, default=TRUE}
}
\value{
a dataframe of concepts and attributes
Is' is a non standard relationship_id
}
\description{
TODO need to sort repeated rows
}
\examples{
omop_relations("Non-invasive blood pressure")
omop_relations("Non-invasive blood pressure",nsteps=2,r_ids=c('Is a','Subsumes'))
#omop_relations("lenalidomide")
#omop_relations(1633308)
#cmde <- omop_relations(v_ids="Cancer Modifier")
# because of R argument matching, you can just use the first unique letters of
# arguments e.g. v for v_ids, cc for cc_ids
chemodrugs <- orels("Cytotoxic chemotherapeutic", v="HemOnc", d="Regimen")
}
