% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_vocabs_preprocess.R
\name{omop_vocabs_preprocess}
\alias{omop_vocabs_preprocess}
\title{read in omop vocabulary csvs downloaded from Athena, preprocess
and write out as binary parquet files for better performance.}
\usage{
omop_vocabs_preprocess(
  from,
  to = tools::R_user_dir("omopcept", which = "cache")
)
}
\arguments{
\item{from}{filepath where omop vocab csvs are stored}

\item{to}{path to save parquet files locally, defaults to package cache which is where omopcept functions look}
}
\value{
nothing
}
\description{
by default saves parquet files locally to package cache where omopcept functions can access
}
\seealso{
\code{\link[=omop_vocab_table_save]{omop_vocab_table_save()}}
alternative that saves an already processed parquet file from remote or local source to local package cache
}
